requirements.txt
opencv-python
numpy
mediapipe
pyserial

